improve(adapters): final OpStack migration
