+++ src/adapter/index.ts
@@ -1,2 +1,3 @@
 export * from "./BaseChainAdapter";
-export { aboveAllowanceThreshold } from "./utils";
+export * from "./AdapterManager";
+export * from "./CrossChainTransferClient";
