--- src/clients/bridges/index.ts
@@ -1,9 +1,0 @@
-export * from "./BaseAdapter";
-export * from "./AdapterManager";
-export * from "./op-stack/OpStackAdapter";
-export * from "./ArbitrumAdapter";
-export * from "./PolygonAdapter";
-export * from "./CrossChainTransferClient";
-export * from "./ZKSyncAdapter";
-export * from "./LineaAdapter";
-export * from "./ScrollAdapter";
