--- src/clients/bridges/op-stack/optimism/index.ts
@@ -1,2 +1,0 @@
-export * from "./DaiOptimismBridge";
-export * from "./SnxOptimismBridge";
