+++ src/monitor/MonitorConfig.ts
@@ -40,9 +40,6 @@ export class MonitorConfig extends CommonConfig {
     token: string;
   }[] = [];
 
-  // TODO: Remove this config once we fully migrate to generic adapters.
-  readonly useGenericAdapter: boolean;
-
   constructor(env: ProcessEnv) {
     super(env);
 
@@ -62,7 +59,6 @@ export class MonitorConfig extends CommonConfig {
       REFILL_BALANCES,
       REFILL_BALANCES_ENABLED,
       STUCK_REBALANCES_ENABLED,
-      MONITOR_USE_GENERIC_ADAPTER,
     } = env;
 
     this.botModes = {
@@ -73,8 +69,6 @@ export class MonitorConfig extends CommonConfig {
       unknownRootBundleCallersEnabled: UNKNOWN_ROOT_BUNDLE_CALLERS_ENABLED === "true",
       stuckRebalancesEnabled: STUCK_REBALANCES_ENABLED === "true",
     };
-
-    this.useGenericAdapter = MONITOR_USE_GENERIC_ADAPTER === "true";
 
     // Used to monitor activities not from whitelisted data workers or relayers.
     this.whitelistedDataworkers = parseAddressesOptional(WHITELISTED_DATA_WORKERS);
