+++ src/relayer/RelayerConfig.ts
@@ -61,9 +61,6 @@ export class RelayerConfig extends CommonConfig {
   // It is up to the user to ensure that the spoke pool on the target chain has tryMulticall in its active implementation.
   readonly tryMulticallChains: number[];
 
-  // TODO: Remove this config item once we fully move to generic chain adapters.
-  readonly useGenericAdapter: boolean;
-
   constructor(env: ProcessEnv) {
     const {
       RELAYER_ORIGIN_CHAINS,
@@ -89,11 +86,8 @@ export class RelayerConfig extends CommonConfig {
       RELAYER_EXTERNAL_INDEXER,
       RELAYER_SPOKEPOOL_INDEXER_PATH,
       RELAYER_TRY_MULTICALL_CHAINS,
-      RELAYER_USE_GENERIC_ADAPTER,
     } = env;
     super(env);
-
-    this.useGenericAdapter = RELAYER_USE_GENERIC_ADAPTER === "true";
 
     // External indexing is dependent on looping mode being configured.
     this.externalIndexer = this.pollingDelay > 0 && RELAYER_EXTERNAL_INDEXER === "true";
