+++ test/mocks/MockCrossChainTransferClient.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from "ethers";
-import { CrossChainTransferClient } from "../../src/clients/bridges";
+import { CrossChainTransferClient } from "../../src/adapter";
 export class MockCrossChainTransferClient extends CrossChainTransferClient {
   crossChainTransferAmount: BigNumber = BigNumber.from(0);
   constructor() {
